Error Free
Managed to get everything fully boilerplated when returning blank strings as payloads in my actions. Now to add functionality, and hope I can find a way to retain state in localstorage. I am guessing this can be done right in the reducer's index.js in a function involving initial state.